requirements.txt
ollama==0.6.1
coqui-tts==0.27.3
torch==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
SpeechRecognition==3.14.4
PyAudio==0.2.14
sounddevice==0.5.2
numpy==1.26.4
deepspeed==0.11.2